Revert "SignatureValidatorV2: must keccak256 px"
Initially I thought it was more secure to hash it, but then figured out
it wasn't. Even in a freakish case that px can be forged, hashing won't
save the situation cause the private key of any EOA is exposed anyway
upon signing. But such a scenario shouldn't be possible, as it
compromises the whole system anyway.

This reverts commit a1d27b44f9b321f7ab08143d90a6347e158be2f2.